[refactor] #108 mapservice to interface
